requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
Django==3.2.3
django-phonenumber-field==5.1.0
gunicorn==20.1.0
idna==2.10
phonenumbers==8.12.23
Pillow==8.2.0
python-decouple==3.4
pytz==2021.1
razorpay==1.2.0
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
waitress==2.0.0
whitenoise==5.2.0
